refactor(xr)!: move createXR/useXR into solid-three/xr
Consolidate all WebXR under the `solid-three/xr` sub-path so core's public surface is XR-free. `createXR`, `useXR`, and the `XRContext`/`XRState` types move out of the package entry; import them from `solid-three/xr` alongside `xrEvents()`.

The sub-path is now a folder mirroring `testing/`: `src/xr/index.tsx` re-exports `src/xr/create-xr.tsx` (session entry) and `src/xr/events.ts` (the controller-events plugin). Core keeps only the renderer-duck-typed frame-loop XR awareness (`gl.xr.isPresenting`), which is loop correctness independent of who drives the session.

Also fixes two latent lint issues the files surface now that they sit under the linted glob: drop an unnecessary optional chain on a non-null controller event, and type `XRRenderer.xr` optional to match its runtime guard (binding it to a local const through the session-start path).
