+++ package.json
@@ -88,6 +88,9 @@
     "*": {
       "testing": [
         "./dist/testing/index.d.ts"
+      ],
+      "xr": [
+        "./dist/xr.d.ts"
       ]
     }
   },
