+++ site/src/routes/api/hooks/create-xr.mdx
@@ -30,7 +30,8 @@ const xr = createXR()
 A VR scene with a DOM "Enter VR" button:
 
 ```tsx
-import { Canvas, createXR } from "solid-three"
+import { Canvas } from "solid-three"
+import { createXR } from "solid-three/xr"
 import { Show } from "solid-js"
 
 function App() {
