+++ src/index.ts
@@ -2,8 +2,6 @@ export { Canvas, type CanvasProps } from "./canvas.tsx"
 export { Entity, Portal, Resource } from "./components.tsx"
 export { $S3C } from "./constants.ts"
 export { createEntity, createT } from "./create-t.tsx"
-export { createXR, useXR } from "./create-xr.tsx"
-export type { XRContext, XRState } from "./create-xr.tsx"
 export { useFrame, useLoader, useThree } from "./hooks.ts"
 export { plugin } from "./plugin.ts"
 export { hasPointerCapture } from "./pointer-capture.ts"
